+++ packages/tools/dom/src/html.js
@@ -4,6 +4,6 @@
 
 export const alignTags = htmlString =>
   htmlString
-  .replace(/>\s*</g, '>\n<')
-  .replace(/>\s*\[/g, '>\n[')
-  .replace(/]\s*</g, ']\n<');
+  .replace(/>\s*</gu, '>\n<')
+  .replace(/>\s*\[/gu, '>\n[')
+  .replace(/\]\s*</gu, ']\n<');