+++ packages/tools/utils/src/[any-number]:accumcb-[array-number].spec.js
@@ -6,13 +6,6 @@ import {arrayMaxWith, arrayMinWith} from "./[any-number]:accumcb-[array-number]"
 
 describe("Function -> (Array -> Number)", function() {
     describe("max/min", function() {
-        const numArray = [-1, -2, 0, 1, 2];
-        const objArray = [
-            {a: -1, b: -1},
-            {a: 0, b: 0},
-            {a: 1, b: 1},
-            {a: 2, b: -2}
-        ];
         describe("arrayMaxWith", function() {
             it("should return a function expecting an array of objects and returning the max of results of applying the provided fuction on all of the array items", function() {
                 const maxWithAbsSin = arrayMaxWith(_.pipe([Math.sin, Math.abs]));