+++ packages/tools/utils/src/any-[array-object].spec.js
@@ -2,10 +2,6 @@ import {strict as assert} from "assert";
 
 import {makeKeyed} from "./any-[array-object]";
 
-function returnArgs () {
-    return arguments;
-}
-
 describe("Any -> (Any -> Array)", function() {
   describe("toFloatOrIdentity", function() {
     it("should return a function expecting an array of keys and returning an object with the provided value as value of those keys", function() {