+++ packages/tools/dev/index.js
@@ -1,2 +1,5 @@
-export * from "./src/common.js"
-export * from "./src/test.js"
+export * from "./src/log/array-[any-any]"
+export * from "./src/log/iterable-array"
+export * from "./src/log/string-[any-any]"
+export * from "./src/common"
+export * from "./src/test"