+++ packages/tools/dev/CHANGELOG.md
@@ -1,12 +1,21 @@
-## `@svizzle/dev` (next)
+## `@svizzle/dev` v0.3.0 (next)
 
 - add linting
 - updated docs
+
+- add `tapMessage`
+- add `tapWith`
+- move `tapAppendTo` from /utils
+- move `tapType` from /utils
+- move `tapTypeAndValue` from /utils
+- move `tapValue` from /utils
+
 
 ## `@svizzle/dev` v0.2.0
 
 - add `renameToExtension`
 - more generic `makeBanner`
+
 
 ## `@svizzle/dev` v0.1.0